% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm.
\code{dijkstra} returns the shortest path between \code{init_node}
and all the other nodes in \code{graph}.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame representing an undirected weighted graph.
The first column fo the data frame v1 lists all the nodes repeated
as many times as they have neighbors. The second column of the data
frame v2 lists per each entry in v1 its neighbor. And the last column
w lists the edge between corresponding entries in v1 and v2. See
\code{wiki_graph} in example.}

\item{init_node}{An integer representing a node in \code{graph}.}
}
\value{
Returns a vector of shortest path distances between
\code{init_node} and all the other nodes in \code{graph}.Each
entry represents the distance between \code{init_node} and the
node that the entry represents.
}
\description{
Dijkstra algorithm.
\code{dijkstra} returns the shortest path between \code{init_node}
and all the other nodes in \code{graph}.
}
\examples{
wiki_graph <-data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
\references{
https://en.wikipedia.org/wiki/Dijkstra's_algorithm
}
