\name{packagelab3-package}
\alias{packagelab3-package}
\alias{packagelab3}
\docType{package}
\title{
packagelab3
}
\description{
\packageDescription{packagelab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{packagelab3}
\packageIndices{packagelab3}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{packagelab3}

Maintainer: \packageMaintainer{packagelab3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
