% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm.
\code{euclidean} returns the GCD of the two arguments.
This function uses the Euclidean algorithm to compute
the GCD of the two arguments.}
\usage{
euclidean(x, y)
}
\arguments{
\item{x, y}{Integer numbers.}
}
\value{
Returns the GCD of x and y as an integer number.
}
\description{
Euclidean algorithm.
\code{euclidean} returns the GCD of the two arguments.
This function uses the Euclidean algorithm to compute
the GCD of the two arguments.
}
\examples{
euclidean(45, 10)
euclidean(10, 5)
\dontrun{
euclidean("b", "a")
}
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
